Finance

+++ Library/Library/Calendar.cs
@@ -9,9 +9,9 @@ namespace Library
     class Calendar
     {
         public DateTime DateOfHire { get; set; }
-        public int Duration { get; set; }
+        public double Duration { get; set; }
         public DateTime EndingDate { get; set; }
-        public Calendar(DateTime dateOfHire, int duration)
+        public Calendar(DateTime dateOfHire, double duration)
         {
             DateOfHire = dateOfHire;
             Duration = duration;
